refactor(spring-data): decompose translator; normalize operand order once
Split the 1,351-line adapter into focused package-private modules with no
public API change:

- Scope.java: variable resolution, relation-chain walking, lambda scoping
- HierarchyTranslator.java: overlaps/ancestorOf/descendentOf + models
- PlanValues.java: proto<->Java conversion, add fold/solve, LIKE escaping
- SpringDataQueryPlanAdapter.java (712 lines): public API + core Translator

Structural cleanups, all behavior-preserving:

- NormalizedBinary states the planner source-order invariant once
  (field-side first, directional ops mirrored on swap), replacing the
  scattered order fixes in the leaf handler, add comparison, size
  comparison, and hasIntersection
- handleIn collapsed from two mirror-image branches into one flow; the
  mapping kind decides membership semantics, not operand order
- correlate() extracted; the Root/Join correlation dance existed twice
- RootScope.resolveMapping now derives from resolveRelationChain;
  duplicate walkRelationChain deleted
- Scope.memberPath() canonicalizes the defaultMemberField fallback that
  lived in three places
- dead Translator.topMapper field and constructor parameter removed

Tests: runCount overloads deduplicated; integration attribute maps now
derived via merge() instead of hand-duplicated entries. 203 tests pass
in the Docker/testcontainers suite.
